pyproject.toml
[project]
name = "neuroice-starter"
version = "0.1.0"
description = "NeuroICE case starter — AI Business SPB Hackathon 2026"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "soundfile>=0.12",
    "scikit-learn>=1.3",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "ipykernel>=7.2.0",
    "pydantic-settings>=2.13.1",
    "fastapi>=0.115",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30",
]

[tool.uv]
package = false
